tests/wpt/web-platform-tests/shadow-dom/declarative/setinnerhtml.tentative.html

<!DOCTYPE html>
<title>getInnerHTML </title>
<link rel='author' title='Mason Freed' href='mailto:masonfreed@chromium.org'>
<link rel='help' href='https://github.com/whatwg/dom/issues/831'>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script src='../resources/shadow-dom-utils.js'></script>

<body>
<script>
function testElementType(allowsShadowDom, elementType, applyToShadow) {
  const t = test(t => {
    // Create and attach element
    let wrapper;
    if (applyToShadow) {
      const host = document.createElement('div');
      t.add_cleanup(function() { host.remove(); });
      document.body.appendChild(host);
      wrapper = host.attachShadow({mode: 'open'});
    } else {
      wrapper = document.createElement('div');
      t.add_cleanup(function() { wrapper.remove(); });
      document.body.appendChild(wrapper);
    }
    const html = `<${elementType}><template shadowroot="open"><slot></slot></template><span></span></${elementType}>`;
    wrapper.setInnerHTML(html, {includeShadowRoots: true});
    if (allowsShadowDom) {
      // Retrieve shadow root
      assert_true(!!wrapper.firstElementChild.shadowRoot,'No shadow root found');
    } else {
      const leftover = wrapper.firstElementChild.firstElementChild;
      assert_true(wrapper.firstElementChild.childElementCount == 0 || leftover instanceof HTMLTemplateElement,'Template should be left over (or no children)');
    }
  }, `${applyToShadow ? 'ShadowRoot' : 'Element'}.setInnerHTML() on <${elementType}>${allowsShadowDom ? `, with declarative Shadow DOM.` : ''}`);
}

function runAllTests() {
  const allElements = [...HTML5_ELEMENT_NAMES, 'htmlunknown'].filter(item => item !== 'body');
  const safelisted = ATTACHSHADOW_SAFELISTED_ELEMENTS;
  for (const elementName of allElements) {
    for (const applyToShadow of [false, true]) {
      testElementType(safelisted.includes(elementName), elementName, applyToShadow);
    }
  }
}

runAllTests();
</script>
